[BugFix] Fix be bugs found by sonar (#14253)

+++ be/src/exec/vectorized/aggregate/aggregate_blocking_node.cpp
@@ -144,7 +144,7 @@ Status AggregateBlockingNode::get_next(RuntimeState* state, ChunkPtr* chunk, boo
 
     // For having
     RETURN_IF_ERROR(ExecNode::eval_conjuncts(_conjunct_ctxs, (*chunk).get()));
-    _aggregator->update_num_rows_returned(-(old_size - (*chunk)->num_rows()));
+    _aggregator->update_num_rows_returned(-(old_size - static_cast<int64_t>((*chunk)->num_rows())));
 
     _aggregator->process_limit(chunk);
